Search Update
Implemented search functionality and error handling for Customer, Address and Appointment. Includes some helper functions to make the process more user-friendly.

+++ CapstoneProject/CustomerScreen.cs
@@ -30,27 +30,43 @@ namespace C969Assessment
             MySqlDataReader reader;
             DataGridViewSelectedRowCollection deleteRows = customerList.SelectedRows;
 
+            // Block deletion if any customer records rely on this address.
+
+
             foreach (DataGridViewRow row in deleteRows)
             {
                 Customer newCust = (Customer)row.DataBoundItem;
-                MySqlCommand command = new MySqlCommand($"DELETE FROM customer WHERE customerId = {newCust.Id}", DatabaseConnection.connection);
-                reader = command.ExecuteReader();
-                reader.Close();
 
-                Customer.custList.Remove(newCust);
+                if (Appointment.getCustAppts(newCust.Id).Count > 0)
+                {
+                    MessageBox.Show("You cannot delete a customer associated with one or more appointments.", "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                else
+                {
+                    MySqlCommand command = new MySqlCommand($"DELETE FROM customer WHERE customerId = {newCust.Id}", DatabaseConnection.connection);
+                    reader = command.ExecuteReader();
+                    reader.Close();
+
+                    Customer.custList.Remove(newCust);
+                }
+
             }
         }
 
         private void refreshBtn_Click(object sender, System.EventArgs e)
         {
             customerList.DataSource = null;
+            Customer.custList = Customer.getCustomers();
             customerList.DataSource = Customer.custList;
+
         }
 
         private void searchBtn_Click(object sender, System.EventArgs e)
         {
-            SearchScreen searchScreen = new SearchScreen(this);
-            searchScreen.Show();
+            CustomerSearch customerSearch = new CustomerSearch();
+            customerSearch.Show();
+            customerSearch.Focus();
         }
     }
 }
